refactor(nursing): update domain models and mapper configurations
- Changed data types from Long to Integer for ID fields in NursingLevel and NursingPlan
- Updated method signatures in mapper interfaces to use Integer instead of Long
- Modified XML mapper files to reflect Integer parameter types
- Removed unused NursingLevelMapper.java file
- Updated Java version in pom.xml from 1.8 to 11
- Moved dashboard view component to correct directory structure
- Added new utility functions for charts and color management
- Introduced constants for dashboard data visualization
- Updated API model descriptions and documentation comments

+++ YiSmartCloud-background/YiSmartCloud-nursing-platform/src/main/java/org/FlyingSparrow/YiSmartCloud/nursing/mapper/NursingLevelMapper.java
@@ -7,7 +7,7 @@ import org.FlyingSparrow.YiSmartCloud.nursing.domain.NursingLevel;
  * 护理等级Mapper接口
  * 
  * @author Siborne
- * @date 2026-02-28
+ * @date 2026-03-01
  */
 public interface NursingLevelMapper 
 {
@@ -17,7 +17,7 @@ public interface NursingLevelMapper
      * @param id 护理等级主键
      * @return 护理等级
      */
-    public NursingLevel selectNursingLevelById(Long id);
+    public NursingLevel selectNursingLevelById(Integer id);
 
     /**
      * 查询护理等级列表
@@ -49,7 +49,7 @@ public interface NursingLevelMapper
      * @param id 护理等级主键
      * @return 结果
      */
-    public int deleteNursingLevelById(Long id);
+    public int deleteNursingLevelById(Integer id);
 
     /**
      * 批量删除护理等级
@@ -57,5 +57,5 @@ public interface NursingLevelMapper
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deleteNursingLevelByIds(Long[] ids);
+    public int deleteNursingLevelByIds(Integer[] ids);
 }
