+++ YiSmartCloud-background/YiSmartCloud-nursing-platform/src/main/java/org/FlyingSparrow/YiSmartCloud/nursing/mapper/NursingPlanMapper.java
@@ -1,13 +1,14 @@
 package org.FlyingSparrow.YiSmartCloud.nursing.mapper;
 
 import java.util.List;
+
 import org.FlyingSparrow.YiSmartCloud.nursing.domain.NursingPlan;
 
 /**
  * 护理计划Mapper接口
  * 
  * @author Siborne
- * @date 2026-02-28
+ * @date 2026-03-01
  */
 public interface NursingPlanMapper 
 {
@@ -17,7 +18,7 @@ public interface NursingPlanMapper
      * @param id 护理计划主键
      * @return 护理计划
      */
-    public NursingPlan selectNursingPlanById(Long id);
+    public NursingPlan selectNursingPlanById(Integer id);
 
     /**
      * 查询护理计划列表
@@ -49,7 +50,7 @@ public interface NursingPlanMapper
      * @param id 护理计划主键
      * @return 结果
      */
-    public int deleteNursingPlanById(Long id);
+    public int deleteNursingPlanById(Integer id);
 
     /**
      * 批量删除护理计划
@@ -57,5 +58,5 @@ public interface NursingPlanMapper
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deleteNursingPlanByIds(Long[] ids);
+    public int deleteNursingPlanByIds(Integer[] ids);
 }
